AI_Agent_Complete/requirements.txt
# 使用清华源加速
--index-url https://pypi.tuna.tsinghua.edu.cn/simple

# LangChain 核心框架与工作流
langchain>=1.2.0
langchain-core>=1.2.5
langchain-community>=0.4.1
langchain-huggingface>=1.2.0
langgraph>=1.0.5

# 向量检索与嵌入
langchain-chroma>=1.1.0
sentence-transformers>=5.2.0
chromadb>=1.4.0

# 文档处理与分词
langchain-text-splitters>=1.1.0
pypdf>=6.5.0
docx2txt>=0.8

# 环境配置与实用工具
python-dotenv>=1.2.1

# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# 配置文件处理
pyyaml>=6.0

# HTTP请求
requests>=2.31.0

# 日期时间处理
python-dateutil>=2.8.0

# 模型加速和分布式推理
accelerate>=0.20.0

# 数据处理
pandas>=1.5.0
numpy>=1.21.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0

# 其他工具
python-multipart>=0.0.6